pyproject.toml
[project]
name = "price-impact"
version = "0.1.0"
description = "Price impact modeling and visualization"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
  "pandas>=2.1,<3",
  "numpy>=1.25,<2",
  "matplotlib>=3.7,<4",
  "seaborn>=0.13,<0.14",
  "pdfminer-six>=20250506",
  "pypdf>=6.0.0",
  "pymupdf>=1.26.4",
  "pytesseract>=0.3.13",
  "pillow>=11.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
